conversion bugfix in RK4 tracker

+++ RK4Tracker.py
@@ -109,7 +109,7 @@ class Combined_RK4:
     
     def update(self, gyro_now, accel_now, dt):
         gyro_now = np.array(gyro_now)
-        accel_now = np.array(accel_now, dtype=float)
+        accel_now = np.array(accel_now)
         
         # 1. Update Orientation (Gyro + Complementary Filter)
         gyro_mid = (self.gyro_prev + gyro_now) / 2.0
